+++ ceammc/ext/src/lib/cicm/Sources/eclass.h
@@ -27,7 +27,6 @@
  */
 
 /*!
- * \fn          t_eclass* eclass_new(const char *name, t_typ_method newm, t_typ_method freem, size_t size, int flags, t_atomtype arg1, int arg2)
  * \brief       Allocates the memory and initialize a new t_eclass.
  * \details     Allocates and intializes the default Pd t_class then enlarges it to fit a t_eclass. \n This function should be replace class_new().
  * \param name  The class name.
@@ -42,7 +41,6 @@
 t_eclass* eclass_new(const char* name, t_typ_method newm, t_typ_method freem, size_t size, int flags, t_atomtype arg1, int arg2);
 
 /*!
- * \fn          void eclass_guiinit(t_eclass* c, long flags)
  * \brief       Initialize the t_eclass for the GUI behavior.
  * \details     Sets the defaults values and initializes the methods of the t_eclass for the GUI behavior.
  * \param c     The t_eclass pointer.
@@ -51,7 +49,6 @@ t_eclass* eclass_new(const char* name, t_typ_method newm, t_typ_method freem, si
 void eclass_guiinit(t_eclass* c, long flags);
 
 /*!
- * \fn          void eclass_init(t_eclass* c, long flags)
  * \brief       Initialize the t_eclass for the GUI behavior (deprecated).
  * \details     Sets the defaults values and initializes the methods of the t_eclass for the GUI behavior.
  * \param c     The t_eclass pointer.
@@ -61,7 +58,6 @@ void eclass_guiinit(t_eclass* c, long flags);
 void eclass_init(t_eclass* c, long flags) _FUNCTION_DEPRECTAED_;
 
 /*!
- * \fn          t_pd_err eclass_register(t_symbol *name, t_eclass *c)
  * \brief       Initialize the t_eclass in a namespace and preprare the t_eclass to be ready.
  * \details     Should always call this function after setting all the methods and attributes.
  * \param name  The namespace obj or gui (dummy).
@@ -70,7 +66,6 @@ void eclass_init(t_eclass* c, long flags) _FUNCTION_DEPRECTAED_;
 t_pd_err eclass_register(t_symbol* name, t_eclass* c);
 
 /*!
- * \fn          void eclass_dspinit(t_eclass* c)
  * \brief       Initialize the t_eclass for the DSP behavior.
  * \details     Sets the defaults values and initializes the methods of the t_eclass for the DSP behavior.
  * \param c     The t_eclass pointer.
@@ -78,7 +73,6 @@ t_pd_err eclass_register(t_symbol* name, t_eclass* c);
 void eclass_dspinit(t_eclass* c);
 
 /*!
- * \fn          void eclass_addmethod(t_eclass* c, method m, const char* name, t_atomtype type, long anything)
  * \brief       Adds a method to the t_eclass.
  * \details     Adds and wraps the methods of the class. For the moment the funtion takes only 1 type the second type is a dummy one and should always be 0.
  * \param c     The t_eclass pointer.
@@ -91,7 +85,6 @@ void eclass_addmethod(t_eclass* c, t_typ_method m, t_symbol* name, t_atomtype ty
 void eclass_addmethod(t_eclass* c, t_typ_method m, const char* name, t_atomtype type, long dummy);
 
 /*!
- * \fn          void eclass_new_attr_typed(t_eclass* c, const char* attrname, const char* type, long size, long maxsize, long flags, long offset)
  * \brief           Creates a new attribute.
  * \details         Allocates the memory and intializes an new attribute for an eclass. You should prefer to use the MACROS.
  * \param c         The t_eclass pointer
@@ -103,9 +96,7 @@ void eclass_addmethod(t_eclass* c, t_typ_method m, const char* name, t_atomtype 
  */
 void eclass_new_attr_typed(t_eclass* c, const char* attrname, const char* type, size_t size, size_t maxsize, size_t offset);
 
-//! @cond
 /*!
- * \fn       void eclass_attr_setter(t_object* x, t_symbol *s, int argc, t_atom *argv)
  * \brief           The setter method of the attributes.
  * \param x         The object pointer
  * \param s         The attribute name
@@ -118,19 +109,17 @@ bool ebox_attr_float_setter(t_ebox* x, t_eattr* a, t_float value, size_t idx, t_
 bool ebox_attr_long_setter(t_ebox* x, t_eattr* a, t_float value, size_t idx, t_eattr_op op);
 
 /*!
- * \fn       int eclass_attr_getter(t_object* x, t_symbol *s, int* argc, t_atom** argv)
  * \brief           The getter method of the attributes.
  * \param x         The object pointer
  * \param s         The attribute name
  * \param argc      The size of the array of atoms
  * \param argv      The array of atoms that contains the attributes values
- * \return 1 on success, 0 on error
+ * \return true on success, false on error
  */
-int eclass_attr_getter(t_object* x, t_symbol* s, int* argc, t_atom** argv);
+bool eclass_attr_getter(t_object* x, t_symbol* s, int* argc, t_atom** argv);
 
 /*!
- * \fn       void eclass_attr_ceammc_getter(t_object* x, t_symbol *s, int* argc, t_atom** argv)
- * \brief           The getter method of the attributes.
+ * \brief           output requested property values to first object outlet
  * \param x         The object pointer
  * \param s         The attribute name
  * \param argc      The size of the array of atoms
@@ -138,10 +127,7 @@ int eclass_attr_getter(t_object* x, t_symbol* s, int* argc, t_atom** argv);
  */
 void eclass_attr_ceammc_getter(t_object* x, t_symbol* s, int argc, t_atom* argv);
 
-//! @endcond
-
 /*!
- * \fn          void eclass_attr_category(t_eclass* c, const char* attrname, long flags, const char* category)
  * \brief           Sets the category of an attribute.
  * \details         For the moment the category is a dummy member of the attribute. You should prefer to use the MACROS.
  * \param c         The t_eclass pointer
@@ -151,7 +137,6 @@ void eclass_attr_ceammc_getter(t_object* x, t_symbol* s, int argc, t_atom* argv)
 void eclass_attr_category(t_eclass* c, const char* attrname, const char* category);
 
 /*!
- * \fn          void eclass_attr_order(t_eclass* c, const char* attrname, long flags, const char* order)
  * \brief           Sets the order of an attribute.
  * \details         For the moment the order is a dummy member of the attribute. You should prefer to use the MACROS.
  * \param c         The t_eclass pointer
@@ -161,7 +146,6 @@ void eclass_attr_category(t_eclass* c, const char* attrname, const char* categor
 void eclass_attr_order(t_eclass* c, const char* attrname, const char* order);
 
 /*!
- * \fn          void eclass_attr_label(t_eclass* c, const char* attrname, long flags, const char* label)
  * \brief           Sets the label of an attribute.
  * \details         The label is the message that will be displayed in the properties window of the object. You should prefer to use the MACROS.
  * \param c         The t_eclass pointer
@@ -171,7 +155,6 @@ void eclass_attr_order(t_eclass* c, const char* attrname, const char* order);
 void eclass_attr_label(t_eclass* c, const char* attrname, const char* label);
 
 /*!
- * \fn          void eclass_attr_style(t_eclass* c, const char* attrname, long flags, const char* style)
  * \brief           Sets the style of an attribute.
  * \details         The style is how the attribute will be displayed in the properties window of the object. it could be \"onoff\", \"color\", \"number\", \"menu\" or \"entry\". You should prefer to use the MACROS.
  * \param c         The t_eclass pointer
@@ -181,7 +164,6 @@ void eclass_attr_label(t_eclass* c, const char* attrname, const char* label);
 void eclass_attr_style(t_eclass* c, const char* attrname, const char* style);
 
 /*!
- * \fn          void eclass_attr_units(t_eclass* c, const char* attrname, long flags, const char* style)
  * \brief           Sets the units of an attribute.
  * \param c         The t_eclass pointer
  * \param attrname  The attribute name
@@ -190,7 +172,6 @@ void eclass_attr_style(t_eclass* c, const char* attrname, const char* style);
 void eclass_attr_units(t_eclass* c, t_symbol* attrname, t_symbol* units);
 
 /*!
- * \fn          void eclass_attr_default(t_eclass* c, const char* attrname, long flags, const char* value)
  * \brief           Sets the default value of an attribute.
  * \details         The default value is a string that will be parse into an array of atoms or a array of number if needed. You should prefer to use the MACROS.
  * \param c         The t_eclass pointer
@@ -200,7 +181,6 @@ void eclass_attr_units(t_eclass* c, t_symbol* attrname, t_symbol* units);
 void eclass_attr_default(t_eclass* c, const char* attrname, const char* value);
 
 /*!
- * \fn          void eclass_attr_filter_min(t_eclass* c, const char* attrname, float value)
  * \brief           Sets the minimum value of an attribute.
  * \details         The minimum value is only valid for number attributes. You should prefer to use the MACROS.
  * \param c         The t_eclass pointer
@@ -210,7 +190,6 @@ void eclass_attr_default(t_eclass* c, const char* attrname, const char* value);
 void eclass_attr_filter_min(t_eclass* c, const char* attrname, t_float value);
 
 /*!
- * \fn          void eclass_attr_filter_max(t_eclass* c, const char* attrname, float value)
  * \brief           Sets the maximum value of an attribute for number attributes.
  * \details         The maximum value is only valid for number attributes. You should prefer to use the MACROS.
  * \param c         The t_eclass pointer
@@ -220,7 +199,6 @@ void eclass_attr_filter_min(t_eclass* c, const char* attrname, t_float value);
 void eclass_attr_filter_max(t_eclass* c, const char* attrname, t_float value);
 
 /*!
- * \fn          void eclass_attr_step(t_eclass* c, const char* attrname, float value)
  * \brief           Sets the step value of an attribute.
  * \details         The step value is only valid for number attributes and will define the step to increment the value in the properties window. You should prefer to use the MACROS.
  * \param c         The t_eclass pointer
@@ -230,7 +208,6 @@ void eclass_attr_filter_max(t_eclass* c, const char* attrname, t_float value);
 void eclass_attr_step(t_eclass* c, const char* attrname, t_float value);
 
 /*!
- * \fn          void eclass_attr_save(t_eclass* c, const char* attrname, long flags)
  * \brief           Sets if the attribute should be saved with the object.
  * \details         Only GUI can save their attributes. You should prefer to use the MACROS.
  * \param c         The t_eclass pointer
@@ -240,7 +217,6 @@ void eclass_attr_step(t_eclass* c, const char* attrname, t_float value);
 void eclass_attr_save(t_eclass* c, const char* attrname, bool value = true);
 
 /*!
- * \fn          void eclass_attr_paint(t_eclass* c, const char* attrname, long flags)
  * \brief           Sets if the attribute should call the ebox_redraw() function.
  * \details         If the method has been called, when the attribute has changed it call the ebox_redraw function to redraw the object. Only GUI can be repainted. You should prefer to use the MACROS.
  * \param c         The t_eclass pointer
@@ -249,7 +225,6 @@ void eclass_attr_save(t_eclass* c, const char* attrname, bool value = true);
 void eclass_attr_paint(t_eclass* c, const char* attrname);
 
 /*!
- * \fn          void eclass_attr_visible(t_eclass* c, const char* attrname, long flags)
  * \brief           Sets if the attribute should be displayed in the properties window.
  * \details         You should prefer to use the MACROS.
  * \param c         The t_eclass pointer
@@ -258,7 +233,6 @@ void eclass_attr_paint(t_eclass* c, const char* attrname);
 void eclass_attr_set_visibility(t_eclass* c, const char* attrname, ceammc::PropValueVis vis);
 
 /*!
- * \fn          void eclass_attr_accessor(t_eclass* c, const char* attrname, t_err_method getter, t_err_method setter)
  * \brief           Initialize new getter and setter methods for the attributes.
  * \details         By default the attribute is initialized with default getter and setter methods. If you just want to be notify when an attribute has changed, you should prefer to use the notify method. You should prefer to use the MACROS.
  * \param c         The t_eclass pointer
@@ -269,7 +243,6 @@ void eclass_attr_set_visibility(t_eclass* c, const char* attrname, ceammc::PropV
 void eclass_attr_accessor(t_eclass* c, const char* attrname, t_getter_method getter, t_setter_method setter);
 
 /*!
- * \fn          void eclass_attr_itemlist(t_eclass* c, const char* attrname, long flags, const char* list)
  * \brief           Sets the list available items of an attribute.
  * \details         For example, the font weight attribute offers \"normal\" and \"bold\". You should prefer to use the MACROS.
  * \param c         The t_eclass pointer