+++ ceammc/CHANGELOG.md
@@ -118,6 +118,7 @@
 - proto.midi.cc - sound_off, notes_off crash fixes
 - expand_env - warning removed while processing @prop requests
 - speech.flite~ - @speed property (now: lesser - slower)
+- ui.tab - 'clear' method renamed to 'erase' preventing name conflict with preset 'clear' method
 
 ## [0.9.6]
 ### Added: